requirements.txt
asttokens==2.2.1
backcall==0.2.0
cbpro==1.1.4
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
importlib-metadata==6.7.0
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pymongo==3.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
requests==2.13.0
six==1.10.0
sortedcontainers==2.4.0
stack-data==0.6.2
-e git+https://github.com/gaussianswan/telegrambot.git@b9eb3377d7fdf5d7b97b72f26f0bec85e9ad4dda#egg=telegrambot
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
wcwidth==0.2.6
websocket-client==0.40.0
zipp==3.15.0
